requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.1
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
bokeh==3.1.1
cached-property==1.5.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cmocean==3.0.3
colorcet==3.0.1
comm==0.1.3
contourpy==1.1.0
cycler==0.11.0
Cython==0.29.36
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
exceptiongroup==1.1.2
executing==1.2.0
fastjsonschema==2.17.1
fonttools==4.40.0
fqdn==1.5.1
frozenlist==1.3.3
homcloud==4.1.0
hopcroftkarp==1.2.5
idna==3.4
imageio==2.31.1
ipycanvas==0.13.1
ipydatawidgets==4.3.5
ipyevents==2.0.1
ipykernel==6.24.0
ipython==8.14.0
ipython-genutils==0.2.0
ipyvtklink==0.2.2
ipywidgets==8.0.7
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.1
jsonpointer==2.4
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_proxy==4.0.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
kiwisolver==1.4.4
lazy_loader==0.3
linkify-it-py==2.0.2
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.8.1
matplotlib-inline==0.1.6
mdit-py-plugins==0.4.0
mdurl==0.1.2
meshio==5.3.4
mistune==3.0.1
msgpack==1.0.5
multidict==6.0.4
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.6.0
nbformat==5.9.0
nest-asyncio==1.5.6
networkx==3.1
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.26.2
opencv-python==4.8.0.74
overrides==7.3.1
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
panel==1.1.1
param==1.13.0
parso==0.8.3
persim==0.3.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.8.0
plotly==5.15.0
pooch==1.7.0
prometheus-client==0.17.0
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
PuLP==2.7.0
pure-eval==0.2.2
pycparser==2.21
pyct==0.5.0
Pygments==2.15.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.7
pythreejs==2.4.2
pytz==2023.3
pyvista==0.40.0
pyviz-comms==2.3.2
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.4.2
ripser==0.6.4
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.11.4
scooby==0.7.2
Send2Trash==1.8.2
simpervisor==1.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
stack-data==0.6.2
tenacity==8.2.2
terminado==0.17.1
threadpoolctl==3.1.0
tifffile==2023.7.10
tinycss2==1.2.1
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
traittypes==0.2.1
trame==2.5.0
trame-client==2.9.4
trame-components==2.1.1
trame-deckgl==2.0.2
trame-markdown==2.0.2
trame-matplotlib==2.0.2
trame-plotly==2.1.1
trame-rca==0.3.1
trame-router==2.0.2
trame-server==2.11.4
trame-simput==2.3.2
trame-vega==2.0.3
trame-vtk==2.5.4
trame-vuetify==2.2.4
typing_extensions==4.7.1
tzdata==2023.3
uc-micro-py==1.0.2
uri-template==1.3.0
urllib3==2.0.3
vtk==9.2.6
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
widgetsnbextension==4.0.8
wrapt==1.15.0
wslink==1.11.1
xyzservices==2023.5.0
yarl==1.9.2
